requirements.txt
discord == 2.2.3
openai>=1.0.0
requests == 2.31.0
python-dotenv == 1.0.0
google-generativeai
PyPDF2
Pillow
python-magic-bin
aiofiles
aiohttp